+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.30.0
 
 * Remove Kafka 3.0.0 and 3.0.1
+* Add support for `simple` authorization and for the User Operator to the `UseKRaft` feature gate 
 * Add network capacity overrides for Cruise Control capacity config
 * The `ServiceAccountPatching` feature gate moves to GA.
   It cannot be disabled anymore and will be permanently enabled.